modified:   include/beg.h 	modified:   include/blackjack.h 	modified:   src/blackjack.c

+++ include/beg.h
@@ -1,7 +1,7 @@
 #ifndef BEG_H
 #define BEG_H
 
-long long Beg(long long *konto);
+long long Beg(int64_t *konto);
 
 #endif // BEG_H
